refactor: move more reusable logic into `api/lib`
