+++ api/items.list.js
@@ -2,7 +2,7 @@
 
 import { getProjectWithItemsQuery } from "./lib/queries.js";
 import { projectFieldsNodesToFieldsMap } from "./lib/project-fields-nodes-to-fields-map.js";
-import { itemFieldsNodesToFieldsMap } from "./lib/item-fields-nodes-to-fields-map.js";
+import { projectItemNodeToGitHubProjectItem } from "./lib/project-item-node-to-github-project-item.js";
 
 /**
  * @param {import("../").default} project
@@ -21,52 +21,7 @@ export default async function listItems(project) {
     projectNext.fields.nodes
   );
 
-  return projectNext.items.nodes.reduce((acc, item) => {
-    const fields = itemFieldsNodesToFieldsMap(
-      projectFields,
-      item.fieldValues.nodes
-    );
-
-    // item is draft
-    if (!item.content) {
-      return acc.concat({
-        id: item.id,
-        fields,
-        isDraft: true,
-      });
-    }
-
-    // item is issue or pull request
-    const common = {
-      id: item.content.id,
-      number: item.content.number,
-      createdAt: item.content.createdAt,
-      closed: item.content.closed,
-      closedAt: item.content.closedAt,
-      assignees: item.content.assignees.nodes.map((node) => node.login),
-      labels: item.content.labels.nodes.map((node) => node.name),
-      repository: item.content.repository.nameWithOwner,
-      milestone: item.content.milestone,
-    };
-    const issueOrPullRequest =
-      item.content.__typename === "Issue"
-        ? {
-            isIssue: true,
-            isPullRequest: false,
-            ...common,
-          }
-        : {
-            isIssue: false,
-            isPullRequest: true,
-            ...common,
-            merged: item.content.merged,
-          };
-
-    return acc.concat({
-      id: item.id,
-      fields,
-      isDraft: false,
-      issueOrPullRequest,
-    });
-  }, []);
+  return projectNext.items.nodes.map((node) =>
+    projectItemNodeToGitHubProjectItem(projectFields, node)
+  );
 }
